extract cycle-related tests to their own file
