+++ src/test.rs
@@ -157,69 +157,6 @@ fn ordering() {
     }
 }
 
-#[test]
-fn cycle_no_solution() {
-    test! {
-        program {
-            trait Foo { }
-            struct S<T> { }
-            impl<T> Foo for S<T> where T: Foo { }
-        }
-
-        // only solution: infinite type S<S<S<...
-        goal {
-            exists<T> {
-                T: Foo
-            }
-        } yields {
-            "No possible solution"
-        }
-    }
-}
-
-#[test]
-fn cycle_many_solutions() {
-    test! {
-        program {
-            trait Foo { }
-            struct S<T> { }
-            struct i32 { }
-            impl<T> Foo for S<T> where T: Foo { }
-            impl Foo for i32 { }
-        }
-
-        // infinite family of solutions: {i32, S<i32>, S<S<i32>>, ... }
-        goal {
-            exists<T> {
-                T: Foo
-            }
-        } yields {
-            "Ambiguous; no inference guidance"
-        }
-    }
-}
-
-#[test]
-fn cycle_unique_solution() {
-    test! {
-        program {
-            trait Foo { }
-            trait Bar { }
-            struct S<T> { }
-            struct i32 { }
-            impl<T> Foo for S<T> where T: Foo, T: Bar { }
-            impl Foo for i32 { }
-        }
-
-        goal {
-            exists<T> {
-                T: Foo
-            }
-        } yields {
-            "Unique; substitution [?0 := i32]"
-        }
-    }
-}
 
 #[test]
 fn multiple_ambiguous_cycles() {
